feat: Split large app.py into modular UI components
- Created transformerlab/ui/ module with specialized components
- config_sidebar.py: Configuration controls and backend selection
- training.py: Model initialization, training, and inference functions
- visualizations.py: Interactive D3.js and Plotly visualizations
- experiments.py: Experiment management and comparison tools
- Reduced main app.py from 820+ lines to 119 lines
- Improved maintainability, testability, and readability
- Each UI module has a focused single responsibility
